api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
neo4j==5.14.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
